requirements.txt
dj-database-url==0.4.2
Django==2.2.7
djangorestframework==3.9.2
gunicorn==20.0.4
psycopg2-binary==2.8.5
whitenoise==5.0.1
djangorestframework-xml==2.0.0